Fixed logging in update method
