+++ app/commands/index.ts
@@ -1,5 +1,7 @@
 export * from './updateFlightChangeData';
+export * from './updateFlightData';
 export * from './updateFlightRegistrationData';
 export * from './updateFlightTimesData';
 export * from './updateFlightTrackData';
+export * from './updateFlightWeatherReports';
 export * from './updateOnTimePerformanceData';
