+++ app/commands/updateFlightTrackData.ts
@@ -1,5 +1,3 @@
-import type { Prisma } from '@prisma/client';
-
 import { fetchFlightTrackData as fetchAdsbExchangeData } from '../data/adsbExchange';
 import { fetchFlightTrackData as fetchPlaneSpottersData } from '../data/planeSpotters';
 import { prisma } from '../db';
@@ -12,15 +10,18 @@ export const updateFlightTrackData = async (
   if (process.env.FLIGHT_TRACKLOG_DATASOURCE === 'flightstats') {
     return;
   }
-  console.log(
-    `Updating flight track data for ${getGroupedFlightsKey(flights[0])}...`,
-  );
+  const flightDataString = getGroupedFlightsKey(flights[0]);
+  console.log(`Fetching flight track data for ${flightDataString}...`);
   const tracklog =
     process.env.FLIGHT_TRACKLOG_DATASOURCE === 'adsbexchange'
       ? await fetchAdsbExchangeData(flights[0])
       : process.env.FLIGHT_TRACKLOG_DATASOURCE === 'planespotters'
         ? await fetchPlaneSpottersData(flights[0])
         : undefined;
+  if (tracklog === undefined) {
+    console.log(`  Flight track data not found for ${flightDataString}.`);
+    return;
+  }
   await prisma.$transaction(
     flights.map(({ id }) =>
       prisma.flight.update({
@@ -28,7 +29,7 @@ export const updateFlightTrackData = async (
           id,
         },
         data: {
-          tracklog: tracklog as Prisma.JsonArray | undefined,
+          tracklog,
         },
       }),
     ),
