+++ app/utils/flights.ts
@@ -21,14 +21,7 @@ import {
 import { toZonedTime } from 'date-fns-tz';
 import groupBy from 'lodash.groupby';
 
-import {
-  updateFlightRegistrationData,
-  updateFlightTimesData,
-  updateFlightTrackData,
-  updateOnTimePerformanceData,
-} from '../commands';
 import { KTS_TO_MPH } from '../commands/constants';
-import type { FlightWithData } from '../commands/types';
 import { SECONDS_IN_HOUR } from '../constants';
 import type { TracklogItem } from '../data/types';
 import {
@@ -50,7 +43,6 @@ import {
   getFlightTimestamps,
 } from './flighttime';
 import { type UserData, fetchGravatarUrl } from './users';
-import { updateFlightWeatherReports } from './weather';
 
 export const flightIncludeObj = {
   user: true,
@@ -191,38 +183,6 @@ const FLIGHT_STATUS_MAP: Record<string, string> = {
   LANDED_TAXIING: 'Landed - Taxiing',
   ARRIVED: 'Arrived',
   CANCELED: '',
-};
-
-export const updateFlightData = async (
-  flights: FlightWithData[],
-): Promise<void> => {
-  let updatedTimesFlights = flights;
-  try {
-    updatedTimesFlights = await updateFlightTimesData(updatedTimesFlights);
-  } catch (err) {
-    console.error(err);
-  }
-  try {
-    updatedTimesFlights =
-      await updateFlightRegistrationData(updatedTimesFlights);
-  } catch (err) {
-    console.error(err);
-  }
-  try {
-    await updateFlightTrackData(updatedTimesFlights);
-  } catch (err) {
-    console.error(err);
-  }
-  try {
-    await updateOnTimePerformanceData(updatedTimesFlights);
-  } catch (err) {
-    console.error(err);
-  }
-  try {
-    await updateFlightWeatherReports(updatedTimesFlights);
-  } catch (err) {
-    console.error(err);
-  }
 };
 
 export const getCenterpoint = (
