+++ app/utils/weather.ts
@@ -1,12 +1,8 @@
-import type { Prisma, WeatherReport } from '@prisma/client';
+import type { WeatherReport } from '@prisma/client';
 import axios, { type AxiosResponse } from 'axios';
-import { fromUnixTime } from 'date-fns';
 import { formatInTimeZone } from 'date-fns-tz';
 
-import type { FlightWithData } from '../commands/types';
-import { getGroupedFlightsKey } from '../commands/utils';
 import { TIMESTAMP_FORMAT_ISO } from '../constants';
-import { prisma } from '../db';
 
 export interface WeatherReportCloudCoverItem {
   [x: string]: string | number | null;
@@ -74,7 +70,7 @@ export const getWeatherReportCloudCoverData = (
 const getUrl = (departureAirportId: string, departureTime: Date): string =>
   `https://aviationweather.gov/api/data/metar?ids=${departureAirportId}&format=json&date=${formatInTimeZone(departureTime, 'UTC', TIMESTAMP_FORMAT_ISO)}`;
 
-const fetchSingleReport = async (
+export const fetchSingleWeatherReport = async (
   airportId: string,
   time: Date,
 ): Promise<AviationWeatherReport | null> => {
@@ -86,119 +82,11 @@ const fetchSingleReport = async (
     departureResult = await axios.get<AviationWeatherReport[]>(
       getUrl(airportId, time),
     );
-    console.log(
-      `  Found ${departureResult.data.length} result${departureResult.data.length !== 1 ? 's' : ''} for ${airportId} at ${time.toISOString()}.`,
-    );
   } catch (err) {
     console.log(
-      `  Unable to fetch weather report for ${airportId} at ${time.toISOString()}.`,
+      `  Weather report not found for ${airportId} at ${time.toISOString()}.`,
     );
     console.error(err);
   }
   return departureResult?.data[0] ?? null;
-};
-
-const getUpdateObject = (
-  data: AviationWeatherReport,
-): Prisma.WeatherReportCreateInput => ({
-  id: data.metar_id,
-  airport: {
-    connect: {
-      id: data.icaoId,
-    },
-  },
-  obsTime: fromUnixTime(data.obsTime),
-  temp: data.temp,
-  dewp: data.dewp,
-  wdir: data.wdir.toString(),
-  wspd: data.wspd,
-  wgst: data.wgst ?? 0,
-  visib: data.visib.toString(),
-  altim: data.altim,
-  wxString: data.wxString,
-  vertVis: data.vertVis,
-  rawOb: data.rawOb,
-  clouds: data.clouds,
-});
-
-export const updateFlightWeatherReports = async (
-  flights: FlightWithData[],
-): Promise<void> => {
-  const {
-    departureAirport,
-    arrivalAirport,
-    diversionAirportId,
-    inTime,
-    inTimeActual,
-    offTime,
-    offTimeActual,
-    onTime,
-    onTimeActual,
-    outTime,
-    outTimeActual,
-  } = flights[0];
-  const departureTime = offTimeActual ?? offTime ?? outTimeActual ?? outTime;
-  const arrivalTime = onTimeActual ?? onTime ?? inTimeActual ?? inTime;
-  const departureWeather: AviationWeatherReport | null =
-    await fetchSingleReport(departureAirport.id, departureTime);
-  const arrivalWeather: AviationWeatherReport | null = await fetchSingleReport(
-    arrivalAirport.id,
-    arrivalTime,
-  );
-  let diversionWeather: AviationWeatherReport | null = null;
-  if (diversionAirportId !== null) {
-    diversionWeather = await fetchSingleReport(diversionAirportId, arrivalTime);
-  }
-  if (
-    departureWeather === null &&
-    arrivalWeather === null &&
-    diversionWeather === null
-  ) {
-    return;
-  }
-  console.log(
-    `Updating weather data for ${getGroupedFlightsKey(flights[0])}...`,
-  );
-  await prisma.$transaction(
-    flights.map(({ id }) =>
-      prisma.flight.update({
-        where: { id },
-        data: {
-          departureWeather:
-            departureWeather !== null
-              ? {
-                  connectOrCreate: {
-                    where: {
-                      id: departureWeather.metar_id,
-                    },
-                    create: getUpdateObject(departureWeather),
-                  },
-                }
-              : undefined,
-          arrivalWeather:
-            arrivalWeather !== null
-              ? {
-                  connectOrCreate: {
-                    where: {
-                      id: arrivalWeather.metar_id,
-                    },
-                    create: getUpdateObject(arrivalWeather),
-                  },
-                }
-              : undefined,
-          diversionWeather:
-            diversionWeather !== null
-              ? {
-                  connectOrCreate: {
-                    where: {
-                      id: diversionWeather.metar_id,
-                    },
-                    create: getUpdateObject(diversionWeather),
-                  },
-                }
-              : undefined,
-        },
-      }),
-    ),
-  );
 };
